Fixed overflow bug

+++ src/it/unimi/dsi/sux4j/mph/MinimalPerfectHashFunction.java
@@ -266,7 +266,7 @@ public class MinimalPerfectHashFunction<T> extends AbstractHashFunction<T> imple
 		
 		if ( n > 0 ) {
 			int m = (int)values.length();
-			count = new int[ ( 2 * m + BITS_PER_BLOCK - 1 ) / BITS_PER_BLOCK ];
+			count = new int[ (int)( ( 2 * (long)m + BITS_PER_BLOCK - 1 ) / BITS_PER_BLOCK ) ];
 			int c = 0;
 
 			final int numWords = ( 2 * m + Long.SIZE - 1 ) / Long.SIZE;
